requirements.txt
aiogram<3.0
environs~=8.0.0
pyzbar~=0.1.9
Pillow~=9.0.1